use vitrual methods instead Interface
