--- src/YesSql.Abstractions/Indexes/IIndexTypeCacheProvider.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Reflection;
-using YesSql.Serialization;
-
-namespace YesSql.Indexes
-{
-    public interface IIndexTypeCacheProvider
-    {
-        public PropertyInfo[] GetTypeProperties(Type type);
-        public PropertyInfoAccessor GetPropertyAccessor(PropertyInfo property);
-    }
-}
