+++ src/YesSql.Core/Configuration.cs
@@ -3,7 +3,6 @@ using Microsoft.Extensions.Logging.Abstractions;
 using System;
 using System.Collections.Generic;
 using System.Data;
-using YesSql.Commands;
 using YesSql.Data;
 using YesSql.Indexes;
 using YesSql.Serialization;
@@ -26,7 +25,7 @@ namespace YesSql
             Logger = NullLogger.Instance;
             ConcurrentTypes = new HashSet<Type>();
             TableNameConvention = new DefaultTableNameConvention();
-            IndexTypeCacheProvider = new DefaultIndexTypeCacheProvider();
+            IndexTypeCacheProvider = new IndexTypeCacheProvider();
         }
 
         public IAccessorFactory IdentifierAccessorFactory { get; set; }
@@ -45,6 +44,6 @@ namespace YesSql
         public ICommandInterpreter CommandInterpreter { get; set; }
         public ISqlDialect SqlDialect { get; set; }
         public IdentityColumnSize IdentityColumnSize { get; set; } = IdentityColumnSize.Int32;
-        public IIndexTypeCacheProvider IndexTypeCacheProvider { get; set; }
+        public IndexTypeCacheProvider IndexTypeCacheProvider { get; set; }
     }
 }
